mofix01
Ignore-this: 2c3622ff4085b21d93fdb997d6c9d663

darcs-hash:20091217140719-68901-90978490fb687c2bfa975d0d1e2afee515ab0a43.gz

+++ kernel/tr/mo/text.cpp
@@ -209,7 +209,7 @@ void deleteTextValue(enum insert_position_t position, xptr node_xptr, strsize_t 
         }
 
         if ((cur_size - size) <= PSTRMAXSIZE) {
-            convertLongtextToText(node_xptr, cur_size - size);
+            convertLongtextToText(node_xptr, (size_t) (cur_size - size));
         }
     }
 }
